+++ lib/src/task_invocation.dart
@@ -5,7 +5,7 @@
 ///
 /// Identifies the [name] of the task to invoke, and any [arguments] to send to
 /// it.
-class TaskInvocation {
+final class TaskInvocation {
   final String name;
   final TaskArgs _arguments;
 
@@ -35,7 +35,7 @@ class TaskInvocation {
 ///
 /// The arguments could be in the form of flags (`grind foo --release`) or flags
 /// (`grind foo --config=bar`).
-class TaskArgs {
+final class TaskArgs {
   final String taskName;
 
   /// The original list of arguments that were parsed.